feat!: change release commands to extension, npm and repo
